Mysterious bug solved. It was a rendering bug.
It was a voxel 3d map rendering bug, it was not rendering
all tiles because the call to the draw function was inside
an if statement for certain walls only or something.

By moving the call to the voxel 3d map draw routine outside
the if statement, the bug was solved, furthermore the tile/voxel
traversal algorithm seems to work good/fine !

Note: the current solution/project is in release mode.